% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/box.cox.R
\name{box.cox}
\alias{box.cox}
\title{box cox transformation}
\usage{
box.cox(x, scale)
}
\arguments{
\item{x}{the variable to transform}

\item{scale}{the scaling power you want to use}
}
\description{
returns a box cox transformed variable,
}
\details{
Detects whether there are zeros in data and adds a constant of 1 if there are
}
\examples{
set.seed(204)
x<-rpois(30, 10)
box.cox(x)

}
