% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minmax.scale.R
\name{minmax.scale}
\alias{minmax.scale}
\title{Min max scaling}
\usage{
minmax.scale(x)
}
\arguments{
\item{x}{the variable to scale}
}
\details{
returns a variable scaled to between 0 and 1
}
\examples{
set.seed(204)
x<-rpois(30, 10)
minmax.scale(x)

}
