% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neater.sumtab.R
\name{neater.sumtab}
\alias{neater.sumtab}
\title{Neater sumtab, summary tables}
\usage{
neater.sumtab(dataset, interest.vars, summary.vars)
}
\arguments{
\item{dataset}{The dataset containing variables to summarize}

\item{interest.vars}{A list of the numeric variables you want summarized}

\item{summary.vars}{A list of factorial variables you want the interest.vars split by}
}
\description{
returns a summary table froma  dataset
}
\details{
Takes a data.frame and summarises the interest.vars by the summary.vars you specify

Outputs a table with columns: 1. Subset, 2 Sample size (N), 3:end summarised variables

Rows: 1. Full data characteristics, 2:end levels of summary.vars

Numeric variables are reported as mean (Sd)
}
\examples{
set.seed(695)
x<-data.frame(gl(2,5,labels=c("Level 1", "Level 2")),rnorm(10,4))
names(x)<-c("factor.var","numeric.var")
neater.sumtab(dataset=x, interest.vars="numeric.var", summary.vars="factor.var")

}
