% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/norm.round.R
\name{norm.round}
\alias{norm.round}
\title{Alternative to round}
\usage{
norm.round(x, digits)
}
\arguments{
\item{x}{the number/variable to round}

\item{digits}{the number of digits to round to}
}
\description{
An alternative round that deals with the fact that 0.5 rounds to either 0 or 1
, the round would always round to 1
}
\details{
a different version of round, shamelessly stolen from here:
https://stackoverflow.com/questions/12688717/round-up-from-5-in-r
this deals with the fact that the standard version of round in r sometimes rounds 0.5
up and sometimes down, for good reasons but reasons that do not apply in some circumstances
}
