% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wapply.R
\name{wapply}
\alias{wapply}
\title{Window apply}
\usage{
wapply(x, width, by = NULL, FUN = NULL, ...)
}
\arguments{
\item{x}{the variable to apply with a window}

\item{width}{window width}

\item{by}{how often to apply the window}

\item{FUN}{function to apply}
}
\description{
Apply across a window,
}
\details{
An apply function for windows, similar to rollapply in the zoo package
}
