% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z.trans.R
\name{z.trans}
\alias{z.trans}
\title{z transformation}
\usage{
z.trans(x)
}
\arguments{
\item{x}{the variable to transform}
}
\details{
returns a z transformed variable, with mean 0 and sd 1
}
\examples{
set.seed(204)
x<-rpois(30, 10)
z.trans(x)

}
